+++ src/animation/programs/mutatePoint.ts
@@ -1,7 +1,12 @@
-import { MutationVector, ShapeDefinition, Vec4 } from "src/lib/types";
 import {
-  isMutationVector,
+  MutationVector,
+  MutationVectorTypes,
+  ShapeDefinition,
+  Vec4,
+} from "src/lib/types";
+import {
   isShapeDefinition,
+  isMutationVector,
   visitShapes,
 } from "src/lib/visit";
 
@@ -9,59 +14,16 @@ export const MAX_MUTATION_VECTORS = 60;
 export const MAX_CONTROLS = 20;
 export const MAX_MUTATION_CONTROL_VECTORS = 120;
 
-export const vectorTypeMapping = {
+export const vectorTypeMapping: Record<MutationVectorTypes, number> = {
   translate: 1,
   stretch: 2,
   rotate: 3,
   deform: 4,
   opacity: 5,
+  lightness: 6,
+  colorize: 7,
+  saturation: 8,
 };
-
-export const mutationControlShader = `
-  uniform vec2 uControlMutationValues[${MAX_MUTATION_CONTROL_VECTORS}];
-  uniform vec3 uMutationValueIndices[${MAX_MUTATION_CONTROL_VECTORS}];
-  uniform vec2 uControlMutationIndices[${MAX_MUTATION_VECTORS}];
-
-  uniform float uControlValues[${MAX_CONTROLS}];
-  uniform vec2 uMutationValues[${MAX_MUTATION_VECTORS}];
-
-  vec2 getMutationValue(int mutationIndex, int mutationType) {
-    vec2 result = uMutationValues[mutationIndex];
-    vec2 controlMutations = uControlMutationIndices[mutationIndex];
-    int start = int(controlMutations.x);
-    int steps = int(controlMutations.y);
-    if (steps == 0) {
-      return result;
-    }
-    for(int i = 0; i < ${MAX_CONTROLS}; i++) {
-      if (i < steps) {
-        vec3 valueIndices = uMutationValueIndices[start + i];
-        // x = offset
-        // y = control value index
-        // z = stepType
-        float controlValue = uControlValues[int(valueIndices.y)];
-
-        int startIndex = int(floor(valueIndices.x + controlValue));
-        int endIndex = int(ceil(valueIndices.x + controlValue));
-        float mixFactor = controlValue - floor(controlValue);
-
-        vec2 mutAValue = uControlMutationValues[startIndex];
-        vec2 mutBValue = uControlMutationValues[endIndex];
-        vec2 mutValue = mix(mutAValue, mutBValue, mixFactor);
-
-        if (mutationType == 2 || mutationType == 5) { // Stretch & Opacity
-          result *= mutValue;
-        } else {
-          result += mutValue;
-        }
-      } else {
-        return result;
-      }
-    }
-
-    return result;
-  }
-`;
 
 export const mutationValueShader = `
   uniform vec2 uMutationValues[${MAX_MUTATION_VECTORS}];
@@ -76,60 +38,78 @@ export const mutationShader = `
 
   // x = type, yz = origin, a = radius
   uniform vec4 uMutationVectors[${MAX_MUTATION_VECTORS}];
-  uniform float uMutationParent[${MAX_MUTATION_VECTORS}];
+  uniform int uMutationParent[${MAX_MUTATION_VECTORS}];
 
-  vec3 mutateOnce(vec3 startValue, int mutationIndex) {
+  mat3 mutateOnce(mat3 startValue, int mutationIndex) {
     vec4 mutation = uMutationVectors[mutationIndex];
     int mutationType = int(mutation.x);
 
     vec2 mutationValue = getMutationValue(mutationIndex, mutationType);
     vec2 origin = mutation.yz;
-    vec3 result = startValue;
+
+    vec3 result = startValue[0];
+    vec3 color = startValue[1];
+    vec3 effect = startValue[2];
 
     if (mutationType == 1) { // Translate
       float effect = 1.0;
-      if (mutation.a > 0.0 && distance(startValue.xy, origin) > mutation.a) {
+      if (mutation.a > 0.0 && distance(result.xy, origin) > mutation.a) {
         effect = 0.0;
       }
-      result = vec3(startValue.xy + mutationValue * effect, startValue.z);
+      result = vec3(result.xy + mutationValue * effect, result.z);
     }
 
     if (mutationType == 2) { // Stretch
       result = vec3(origin.xy + vec2(
-        (startValue.x - origin.x) * mutationValue.x, 
-        (startValue.y - origin.y) * mutationValue.y
-      ), startValue.z);
+        (result.x - origin.x) * mutationValue.x,
+        (result.y - origin.y) * mutationValue.y
+      ), result.z);
     }
 
     if (mutationType == 3) { // Rotation
       float rotation = mutationValue.x * PI_FRAC;
       mat2 entityRotationMatrix = mat2(cos(rotation), sin(rotation), -sin(rotation), cos(rotation));
-      result = vec3((startValue.xy - origin) * entityRotationMatrix + origin, startValue.z);
+      result = vec3((result.xy - origin) * entityRotationMatrix + origin, result.z);
     }
 
     if (mutationType == 4) { // Deform
-      float effect = 1.0 - clamp(distance(startValue.xy, origin), 0.0, mutation.a) / mutation.a;	
-      result = vec3(startValue.xy + mutationValue * effect, startValue.z);	
+      float effect = 1.0 - clamp(distance(result.xy, origin), 0.0, mutation.a) / mutation.a;
+      result = vec3(result.xy + mutationValue * effect, result.z);
     }
 
     if (mutationType == 5) { // Opacity
-      float opacity = mutationValue.x;
-      result = vec3(startValue.xy, startValue.z * opacity);	
+      result = vec3(result.xy, result.z * mutationValue.x);
     }
 
-    return result;
+    if (mutationType == 6) { // Lightness
+      color = vec3(mutationValue.x * color.x, color.yz);
+    }
+
+    if (mutationType == 7) { // Colorize setting
+      effect = vec3(mutationValue.xy, effect.z);
+    }
+
+    if (mutationType == 8) { // Saturation
+      color = vec3(color.x, mutationValue.x * color.y, color.z);
+    }
+
+    return mat3(
+      result,
+      color,
+      effect
+    );
   }
 
-  vec3 mutatePoint(vec3 startValue, int mutationIndex) {
+  mat3 mutatePoint(mat3 startValue, int mutationIndex) {
     int currentNode = mutationIndex;
-    vec3 result = startValue;
+    mat3 result = startValue;
 
     for(int i = 0; i < ${MAX_MUTATION_VECTORS}; i++) {
         if (currentNode == -1) {
             return result;
         }
         result = mutateOnce(result, currentNode);
-        currentNode = int(uMutationParent[currentNode]);
+        currentNode = uMutationParent[currentNode];
     }
     return result;
   }
@@ -169,7 +149,7 @@ const getParentMutation = (
   return null;
 };
 
-export const createMutationList = (
+export const createShapeMutationList = (
   shapes: ShapeDefinition[]
 ): {
   parentList: Float32Array;