+++ src/screens/Composition.tsx
@@ -6,6 +6,7 @@ import React, {
   useMemo,
   useState,
 } from "react";
+import LayerMouseControl from "src/components/LayerMouseControl";
 import MenuItem from "src/components/MenuItem";
 import { ToolbarLabel, ToolbarSpacer } from "src/components/Toolbar";
 import ToolbarMeter from "src/components/ToolbarMeter";
@@ -13,12 +14,12 @@ import {
   combineKeyFrames,
   defaultValueForVector,
   mixVec2,
+  mixHueVec2,
 } from "src/lib/vertices";
-import { isMutationVector, visit } from "src/lib/visit";
-import { maxZoomFactor } from "src/lib/webgl";
+import { isMutationVector, isShapeMutationVector, visit } from "src/lib/visit";
 import CompositionCanvas from "../animation/CompositionCanvas";
 import Menu from "../components/Menu";
-import MouseControl, { MouseMode } from "../components/MouseControl";
+import { MouseMode } from "../components/MouseControl";
 import ShapeList from "../components/ShapeList";
 import ToolbarButton from "../components/ToolbarButton";
 import {
@@ -106,16 +107,10 @@ const Composition: React.VFC<CompositionProps> = ({
   panXState,
   panYState,
 }) => {
-  const [zoom, setZoom] = zoomState;
-  const [panX, setPanX] = panXState;
-  const [panY, setPanY] = panYState;
-  const [isMouseDown, setIsMouseDown] = useState<false | [number, number]>(
-    false
-  );
-  const [mouseMoveDelta, setMouseMoveDelta] = useState<[number, number]>([
-    0,
-    0,
-  ]);
+  const [zoom] = zoomState;
+  const [panX] = panXState;
+  const [panY] = panYState;
+
   const [layerSelected, setLayerSelected] = useState<null | ItemSelection>(
     null
   );
@@ -224,33 +219,17 @@ const Composition: React.VFC<CompositionProps> = ({
         control.steps[endStep][vectorKey] ||
         defaultValueForVector(vectorMapping[vectorKey].type);
 
-      const vectorValue = mixVec2(min, max, mixValue);
+      const vectorValue =
+        vectorMapping[vectorKey].type === "colorize"
+          ? mixHueVec2(min, max, mixValue)
+          : mixVec2(min, max, mixValue);
       return { ...result, [vectorKey]: vectorValue };
     }, {} as Keyframe);
     return combineKeyFrames(result, mixed, vectorMapping);
   }, defaultFrameValues);
 
-  const mouseDown = (event: React.MouseEvent) => {
-    const canvasPos = event.currentTarget.getBoundingClientRect();
-    const elementX = event.pageX - canvasPos.left;
-    const elementY = event.pageY - canvasPos.top;
-    setIsMouseDown([elementX, elementY]);
-    setMouseMoveDelta([0, 0]);
-  };
-
-  const mouseMove = (event: React.MouseEvent) => {
-    if (isMouseDown) {
-      const canvasPos = event.currentTarget.getBoundingClientRect();
-      const elementX = event.pageX - canvasPos.left;
-      const elementY = event.pageY - canvasPos.top;
-
-      const deltaX = elementX - isMouseDown[0];
-      const deltaY = elementY - isMouseDown[1];
-      setMouseMoveDelta([deltaX, deltaY]);
-
-      const moveDeltaX = (deltaX - mouseMoveDelta[0]) / zoom;
-      const moveDeltaY = (deltaY - mouseMoveDelta[1]) / zoom;
-
+  const handleDrag = useCallback(
+    (event: React.MouseEvent, deltaX: number, deltaY: number) => {
       if ((shapeSelected || vectorSelected) && event.shiftKey) {
         updateImageDefinition((image) =>
           visit(image, (item, parents) => {
@@ -267,13 +246,13 @@ const Composition: React.VFC<CompositionProps> = ({
               return {
                 ...item,
                 translate: [
-                  Math.round(item.translate[0] + moveDeltaX),
-                  Math.round(item.translate[1] + moveDeltaY),
+                  Math.round(item.translate[0] + deltaX),
+                  Math.round(item.translate[1] + deltaY),
                 ],
               };
             }
             if (
-              isMutationVector(item) &&
+              isShapeMutationVector(item) &&
               ((vectorSelected && item.name === vectorSelected.name) ||
                 (shapeSelected &&
                   parents.find((e) => e.name === shapeSelected.name)))
@@ -281,54 +260,20 @@ const Composition: React.VFC<CompositionProps> = ({
               return {
                 ...item,
                 origin: [
-                  Math.round(item.origin[0] + moveDeltaX),
-                  Math.round(item.origin[1] + moveDeltaY),
+                  Math.round(item.origin[0] + deltaX),
+                  Math.round(item.origin[1] + deltaY),
                 ],
               };
             }
             return undefined;
           })
         );
-        return;
+        return true;
       }
-      const newPanX = Math.min(
-        1.0,
-        Math.max(
-          panX +
-            (((deltaX - mouseMoveDelta[0]) / canvasPos.width) *
-              window.devicePixelRatio) /
-              zoom,
-          -1.0
-        )
-      );
-      setPanX(newPanX);
-
-      const newPanY = Math.min(
-        1.0,
-        Math.max(
-          panY +
-            (((deltaY - mouseMoveDelta[1]) / canvasPos.height) *
-              window.devicePixelRatio *
-              -1.0) /
-              zoom,
-          -1.0
-        )
-      );
-      setPanY(newPanY);
-    }
-  };
-
-  const mouseUp = () => {
-    setIsMouseDown(false);
-  };
-
-  const mouseWheel = (delta: number) => {
-    const z = Math.min(
-      maxZoomFactor(texture),
-      Math.max(0.1, zoom - delta / 100)
-    );
-    setZoom(z);
-  };
+      return false;
+    },
+    [updateImageDefinition, shapeSelected, vectorSelected]
+  );
 
   return (
     <ScreenLayout
@@ -722,12 +667,13 @@ const Composition: React.VFC<CompositionProps> = ({
         ),
       ]}
       main={
-        <MouseControl
+        <LayerMouseControl
           mode={mouseMode}
-          onMouseDown={mouseDown}
-          onMouseMove={mouseMove}
-          onMouseUp={mouseUp}
-          onWheel={mouseWheel}
+          handleDrag={handleDrag}
+          texture={texture}
+          zoomState={zoomState}
+          panXState={panXState}
+          panYState={panYState}
         >
           <CompositionCanvas
             image={texture}
@@ -739,7 +685,7 @@ const Composition: React.VFC<CompositionProps> = ({
             activeLayer={layerSelected}
             showFPS={showFPS}
           />
-        </MouseControl>
+        </LayerMouseControl>
       }
     />
   );