+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-11-17
+
+### Added
+
+- Added `Lightness` mutation to make colors darker / lighter.
+- Added `Color Transition` mutation to transition color to target color.
+- Added `Target Color` mutation as target for Colorize.
+- Added option to reload texture from disk.
+
+### Fixed
+
+- Minor issue with colors in light mode
+
+### Changed
+
+- Updated Electron to 15.2.0
+
 ## [1.2.0] - 2021-05-05
 
 ### Added